backend/app/pyproject.toml
[tool.poetry]
name = "classes-to-calendar-backend"
version = "0.0.1"
description = "FastAPI backend for the Classes to Calendar application"
authors = ["Broden Wanner <brodenwanner@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.0"
python-multipart = "^0.0.5"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
pydantic = {extras = ["email"], version = "^1.8.2"}
beautifulsoup4 = "^4.9.3"
python-dotenv = "^0.19.0"
isort = "^5.9.3"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
mypy = "^0.910"
black = "^21.7b0"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"

[tool.isort]
profile = "black"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
